Refactor WGBS and ATAC
