+++ src/app/omes/(multiomics)/ATAC/downloads/page.tsx
@@ -1,46 +1,29 @@
 "use client";
 import { useATACData } from "@/common/hooks/omeHooks/useATACData";
-import ATACDownloadsTable from "./ATACDownloadsTable";
-import { Protocol, Sex, Site, Status } from "@/common/types/globalTypes";
-import OmeDownloadLayout from "@/common/components/Downloads/OmeDownloadLayout";
-import { useOmeDownloadFiles } from "@/common/hooks/useOmeDownloadFiles";
 import { OmeEnum } from "@/common/types/generated/graphql";
+import OmeDualPaneDownloads from "@/common/components/Downloads/OmeDualPaneDownloads";
+import type { OmeDownloadsConfig } from "@/common/components/Downloads/types";
 
-const ATACDescriptions = [
-  "Fold change signal",
-  "p-value signal",
-  "FDR 0.05 peaks",
-  "Pseudorep peaks",
-];
-
-const ATACDownloads = () => {
-  const ATACData = useATACData({ skip: false });
-  const { data: downloadFiles, loading } = useOmeDownloadFiles(OmeEnum.AtacSeq);
-
-  const rows = ATACData.data ?? [];
-
-  return (
-    <OmeDownloadLayout
-      rows={rows}
-      downloadFiles={downloadFiles ?? []}
-      descriptions={ATACDescriptions}
-      includeProtocolFilter
-      getFilterFields={(row) => ({
-        site: row.site as Site,
-        status: row.status as Status,
-        sex: row.sex as Sex,
-        protocol: row.protocol?.replace(" method", "") as Protocol,
-      })}
-      renderTable={(filteredRows, filteredDownloadFiles) => (
-        <ATACDownloadsTable 
-          rows={filteredRows}
-          ATACData={ATACData} 
-          files={filteredDownloadFiles} 
-          loadingFiles={loading}
-        />
-      )}
-    />
-  );
+type ATACRow = {
+  sample_id: string;
+  site: string;
+  status: string;
+  sex: string;
+  protocol: string;
+  opc_id: string;
 };
+
+const config: OmeDownloadsConfig<ATACRow> = {
+  ome: OmeEnum.AtacSeq,
+  useData: () => useATACData({ skip: false }),
+  datasetFilters: [
+    { field: "sex", label: "Sex", type: "checkbox" },
+    { field: "status", label: "Status", type: "checkbox" },
+    { field: "protocol", label: "Protocol", type: "checkbox" },
+    { field: "site", label: "Site", type: "multiselect" },
+  ],
+};
+
+const ATACDownloads = () => <OmeDualPaneDownloads config={config} />;
 
 export default ATACDownloads;
